requirements.txt
altair==4.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
coverage==6.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.0.0rc8
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.31.2
gitdb==4.0.9
gitdb2==2.0.6
GitPython==2.1.15
honcho==1.1.0
htmlmin==0.1.12
hypothesis==6.54.3
idna==3.3
ImageHash==4.2.1
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==4.11.4
importlib-resources==5.7.1
iniconfig==1.1.1
ipykernel==6.13.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
missingno==0.5.1
mistune==0.8.4
multimethod==1.8
multitasking==0.0.10
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
notebook==6.4.12
npm==0.1.1
numpy==1.22.3
optional-django==0.1.0
packaging==21.3
pandas==1.4.2
pandas-profiling==3.2.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.1.0
plotly==5.8.1
plotly-express==0.4.1
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==8.0.0
pycparser==2.21
pydantic==1.9.1
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.1.2
pytest-arraydiff==0.5.0
pytest-astropy==0.10.0
pytest-astropy-header==0.2.1
pytest-cov==3.0.0
pytest-doctestplus==0.12.0
pytest-filter-subpackage==0.1.1
pytest-mock==3.8.2
pytest-openfiles==0.5.0
pytest-remotedata==0.3.3
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.1.0
requests==2.28.0
rich==12.4.4
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
semantic-version==2.8.5
semver==2.13.0
Send2Trash==1.8.0
six==1.15.0
smmap==5.0.0
smmap2==3.0.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.2.0
statsmodels==0.13.2
streamlit==1.11.1
tangled-up-in-unicode==0.2.0
tenacity==8.0.1
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.20.0
visions==0.7.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
yfinance==0.1.70
zipp==3.8.0